Clean up imports and whitespace.

+++ src/test/java/com/google/cloud/genomics/dataflow/functions/verifybamid/ReadFunctionsTest.java
@@ -188,7 +188,7 @@ public class ReadFunctionsTest {
             alignedRef.substring(i, i + 1),
             Position.newBuilder()
             .setReferenceName(chromosome)
-            .setPosition((long) (originalPos + refPositionOffset))
+            .setPosition(originalPos + refPositionOffset)
             .build());
         qualityOffset++;
         refPositionOffset++;
